+++ commands/deployments/workflows/dep_wf_exec.go
@@ -15,6 +15,7 @@
 package workflows
 
 import (
+	"bytes"
 	"fmt"
 	"net/http"
 	"path"
@@ -31,6 +32,7 @@ func init() {
 	var shouldStreamEvents bool
 	var continueOnError bool
 	var workflowName string
+	var jsonParam string
 	var wfExecCmd = &cobra.Command{
 		Use:     "execute <id>",
 		Short:   "Trigger a custom workflow on deployment <id>",
@@ -50,7 +52,12 @@ func init() {
 			if continueOnError {
 				url = url + "?continueOnError"
 			}
-			request, err := client.NewRequest("POST", url, nil)
+			var request *http.Request
+			if len(jsonParam) == 0 {
+				request, err = client.NewRequest("POST", url, nil)
+			} else {
+				request, err = client.NewRequest("POST", url, bytes.NewBuffer([]byte(jsonParam)))
+			}
 			if err != nil {
 				httputil.ErrExit(err)
 			}
@@ -74,8 +81,9 @@ func init() {
 			return nil
 		},
 	}
-	wfExecCmd.PersistentFlags().StringVarP(&workflowName, "workflow-name", "w", "", "The workflows name")
+	wfExecCmd.PersistentFlags().StringVarP(&workflowName, "workflow-name", "w", "", "The workflows name (mandatory)")
 	wfExecCmd.PersistentFlags().BoolVarP(&continueOnError, "continue-on-error", "", false, "By default if an error occurs in a step of a workflow then other running steps are cancelled and the workflow is stopped. This flag allows to continue to the next steps even if an error occurs.")
+	wfExecCmd.PersistentFlags().StringVarP(&jsonParam, "data", "d", "", "Provide the JSON format for the nodes selection")
 	wfExecCmd.PersistentFlags().BoolVarP(&shouldStreamLogs, "stream-logs", "l", false, "Stream logs after triggering a workflow. In this mode logs can't be filtered, to use this feature see the \"log\" command.")
 	wfExecCmd.PersistentFlags().BoolVarP(&shouldStreamEvents, "stream-events", "e", false, "Stream events after triggering a workflow.")
 	workflowsCmd.AddCommand(wfExecCmd)